fix: firebase config file
